requirements.txt
chromadb
langchain
langchain-anthropic
numpy
langchain-huggingface 
sentence-transformers
python-dotenv